Migrations

+++ database/migrations/2024_08_11_035506_create_orders_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained();
-            $table->float('total_amount');
+            $table->decimal('total_amount', 16, 2);
             $table->enum('status', ['pending', 'completed', 'cancelled']);
             $table->text('shipping_address');
             $table->enum('payment_status', ['unpaid', 'paid', 'refunded']);
