+++ database/migrations/2024_08_11_035713_create_order_items_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('order_id')->constrained();
             $table->foreignId('product_id')->constrained();
             $table->integer('quantity');
-            $table->float('price');
+            $table->decimal('price', 16, 2);
             $table->timestamps();
         });
     }
